src/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
drf-yasg==1.20.0
djoser==2.1.0
